+++ app/src/main/AndroidManifest.xml
@@ -6,6 +6,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:usesCleartextTraffic="true"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -17,14 +18,14 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".ui.EditImageActivity"
+            android:name=".ui.editimage.EditImageActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.EditContentActivity"
+            android:name=".ui.eidtcontent.EditContentActivity"
             android:windowSoftInputMode="adjustNothing"
             android:exported="false" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
