+++ app/src/main/java/org/sopt/zepeto/ui/editimage/EditImageActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.zepeto.ui
+package org.sopt.zepeto.ui.editimage
 
 import android.content.Intent
 import android.net.Uri
@@ -7,6 +7,7 @@ import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import org.sopt.zepeto.databinding.ActivityEditImageBinding
+import org.sopt.zepeto.ui.eidtcontent.EditContentActivity
 
 class EditImageActivity : AppCompatActivity() {
     private lateinit var binding: ActivityEditImageBinding
