+++ app/src/main/java/org/sopt/zepeto/ui/eidtcontent/EditContentActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.zepeto.ui
+package org.sopt.zepeto.ui.eidtcontent
 
 import android.content.Context
 import android.graphics.Rect
@@ -11,7 +11,8 @@ import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import org.sopt.zepeto.databinding.ActivityEditContentBinding
-import org.sopt.zepeto.util.ViewPagerAdapter
+import org.sopt.zepeto.ui.editimage.EditImageActivity
+import org.sopt.zepeto.ui.main.MainActivity
 import org.sopt.zepeto.util.ZepetoSnackBar
 
 class EditContentActivity : AppCompatActivity() {
