+++ app/src/main/java/org/sopt/zepeto/ui/main/ViewPagerAdapter.kt
@@ -1,10 +1,14 @@
-package org.sopt.zepeto.util
+package org.sopt.zepeto.ui.main
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import org.sopt.zepeto.ui.MainActivity
-import org.sopt.zepeto.ui.main.*
+import org.sopt.zepeto.ui.main.MainActivity
+import org.sopt.zepeto.ui.main.feed.FeedFragment
+import org.sopt.zepeto.ui.main.home.HomeFragment
+import org.sopt.zepeto.ui.main.make.MakeFragment
+import org.sopt.zepeto.ui.main.profile.ProfileFragment
+import org.sopt.zepeto.ui.main.world.WorldFragment
 
 class ViewPagerAdapter(fragment: FragmentActivity) : FragmentStateAdapter(fragment) {
     override fun getItemCount() = 5
