+++ app/src/main/java/org/sopt/zepeto/ui/main/profile/ProfileFragment.kt
@@ -1,21 +1,21 @@
-package org.sopt.zepeto.ui.main
+package org.sopt.zepeto.ui.main.profile
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.sopt.zepeto.databinding.FragmentHomeBinding
+import org.sopt.zepeto.databinding.FragmentProfileBinding
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
+class ProfileFragment: Fragment() {
+    private var _binding: FragmentProfileBinding? = null
     private val binding get() = _binding ?: error("Binding이 초기화 되지 않았습니다")
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentHomeBinding.inflate(layoutInflater, container, false)
+        _binding = FragmentProfileBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
 
@@ -23,4 +23,5 @@ class HomeFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+
 }
