+++ app/src/main/res/layout/activity_edit_content.xml
@@ -18,7 +18,7 @@
         android:layout_height="match_parent"
         android:focusable="true"
         android:focusableInTouchMode="true"
-        tools:context=".ui.EditContentActivity">
+        tools:context=".ui.eidtcontent.EditContentActivity">
 
         <View
             android:id="@+id/view_full_cover"
