+++ app/src/main/res/layout/activity_edit_image.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".ui.EditImageActivity">
+    tools:context=".ui.editimage.EditImageActivity">
 
     <ImageView
         android:id="@+id/iv_edit_image"
