+++ app/src/main/res/layout/fragment_profile.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.ProfileFragment">
+    tools:context=".ui.main.profile.ProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
