+++ app/src/main/res/layout/fragment_world.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.WorldFragment">
+    tools:context=".ui.main.world.WorldFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
